required changes to make backend work
